python/requirements.txt
# 飞书 API SDK
lark-oapi>=1.4.15
# Redis 支持
redis>=6.4.0
# FastAPI 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
# 数据验证（pydantic v2 + settings 支持）
pydantic>=2.6
pydantic-settings>=2.2


# 其他依赖
requests>=2.25.0
PyYAML>=5.4.0